+++ lib/axn/webhooks.rb
@@ -8,6 +8,8 @@ require_relative "webhooks/request"
 require_relative "webhooks/signature"
 require_relative "webhooks/resolvers"
 require_relative "webhooks/verify"
+require_relative "webhooks/verifiers"
+require_relative "webhooks/inbound"
 
 module Axn
   module Webhooks